+++ HL7-dotnetcore.sln
@@ -1,11 +1,10 @@
-
 Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio 15
-VisualStudioVersion = 15.0.26124.0
+# Visual Studio Version 16
+VisualStudioVersion = 16.0.28922.388
 MinimumVisualStudioVersion = 15.0.26124.0
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "hl7-dotnetcore", "src\HL7-dotnetcore.csproj", "{8D0228F0-7BE1-4D45-9EB4-8E170F76A17E}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "HL7-dotnetcore", "src\HL7-dotnetcore.csproj", "{8D0228F0-7BE1-4D45-9EB4-8E170F76A17E}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "hl7test", "test\HL7test.csproj", "{910F6F49-6AD7-45EF-8FA0-E141C77D81DA}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "HL7test", "test\HL7test.csproj", "{910F6F49-6AD7-45EF-8FA0-E141C77D81DA}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -16,33 +15,36 @@ Global
 		Release|x64 = Release|x64
 		Release|x86 = Release|x86
 	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
 		{8D0228F0-7BE1-4D45-9EB4-8E170F76A17E}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{8D0228F0-7BE1-4D45-9EB4-8E170F76A17E}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{8D0228F0-7BE1-4D45-9EB4-8E170F76A17E}.Debug|x64.ActiveCfg = Debug|x64
-		{8D0228F0-7BE1-4D45-9EB4-8E170F76A17E}.Debug|x64.Build.0 = Debug|x64
-		{8D0228F0-7BE1-4D45-9EB4-8E170F76A17E}.Debug|x86.ActiveCfg = Debug|x86
-		{8D0228F0-7BE1-4D45-9EB4-8E170F76A17E}.Debug|x86.Build.0 = Debug|x86
+		{8D0228F0-7BE1-4D45-9EB4-8E170F76A17E}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{8D0228F0-7BE1-4D45-9EB4-8E170F76A17E}.Debug|x64.Build.0 = Debug|Any CPU
+		{8D0228F0-7BE1-4D45-9EB4-8E170F76A17E}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{8D0228F0-7BE1-4D45-9EB4-8E170F76A17E}.Debug|x86.Build.0 = Debug|Any CPU
 		{8D0228F0-7BE1-4D45-9EB4-8E170F76A17E}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{8D0228F0-7BE1-4D45-9EB4-8E170F76A17E}.Release|Any CPU.Build.0 = Release|Any CPU
-		{8D0228F0-7BE1-4D45-9EB4-8E170F76A17E}.Release|x64.ActiveCfg = Release|x64
-		{8D0228F0-7BE1-4D45-9EB4-8E170F76A17E}.Release|x64.Build.0 = Release|x64
-		{8D0228F0-7BE1-4D45-9EB4-8E170F76A17E}.Release|x86.ActiveCfg = Release|x86
-		{8D0228F0-7BE1-4D45-9EB4-8E170F76A17E}.Release|x86.Build.0 = Release|x86
+		{8D0228F0-7BE1-4D45-9EB4-8E170F76A17E}.Release|x64.ActiveCfg = Release|Any CPU
+		{8D0228F0-7BE1-4D45-9EB4-8E170F76A17E}.Release|x64.Build.0 = Release|Any CPU
+		{8D0228F0-7BE1-4D45-9EB4-8E170F76A17E}.Release|x86.ActiveCfg = Release|Any CPU
+		{8D0228F0-7BE1-4D45-9EB4-8E170F76A17E}.Release|x86.Build.0 = Release|Any CPU
 		{910F6F49-6AD7-45EF-8FA0-E141C77D81DA}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{910F6F49-6AD7-45EF-8FA0-E141C77D81DA}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{910F6F49-6AD7-45EF-8FA0-E141C77D81DA}.Debug|x64.ActiveCfg = Debug|x64
-		{910F6F49-6AD7-45EF-8FA0-E141C77D81DA}.Debug|x64.Build.0 = Debug|x64
-		{910F6F49-6AD7-45EF-8FA0-E141C77D81DA}.Debug|x86.ActiveCfg = Debug|x86
-		{910F6F49-6AD7-45EF-8FA0-E141C77D81DA}.Debug|x86.Build.0 = Debug|x86
+		{910F6F49-6AD7-45EF-8FA0-E141C77D81DA}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{910F6F49-6AD7-45EF-8FA0-E141C77D81DA}.Debug|x64.Build.0 = Debug|Any CPU
+		{910F6F49-6AD7-45EF-8FA0-E141C77D81DA}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{910F6F49-6AD7-45EF-8FA0-E141C77D81DA}.Debug|x86.Build.0 = Debug|Any CPU
 		{910F6F49-6AD7-45EF-8FA0-E141C77D81DA}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{910F6F49-6AD7-45EF-8FA0-E141C77D81DA}.Release|Any CPU.Build.0 = Release|Any CPU
-		{910F6F49-6AD7-45EF-8FA0-E141C77D81DA}.Release|x64.ActiveCfg = Release|x64
-		{910F6F49-6AD7-45EF-8FA0-E141C77D81DA}.Release|x64.Build.0 = Release|x64
-		{910F6F49-6AD7-45EF-8FA0-E141C77D81DA}.Release|x86.ActiveCfg = Release|x86
-		{910F6F49-6AD7-45EF-8FA0-E141C77D81DA}.Release|x86.Build.0 = Release|x86
+		{910F6F49-6AD7-45EF-8FA0-E141C77D81DA}.Release|x64.ActiveCfg = Release|Any CPU
+		{910F6F49-6AD7-45EF-8FA0-E141C77D81DA}.Release|x64.Build.0 = Release|Any CPU
+		{910F6F49-6AD7-45EF-8FA0-E141C77D81DA}.Release|x86.ActiveCfg = Release|Any CPU
+		{910F6F49-6AD7-45EF-8FA0-E141C77D81DA}.Release|x86.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {055131AB-67EC-486A-8ED8-69B4EBFC0D79}
 	EndGlobalSection
 EndGlobal
