+++ test/Hl7Test.cs
@@ -1,0 +1,299 @@
+using System;
+using System.IO;
+using System.Reflection;
+using HL7.Dotnetcore;
+using Xunit;
+
+namespace HL7.Dotnetcore.Test
+{
+    public class HL7Test
+    {
+        private string HL7_ORM;
+        private string HL7_ADT;
+
+        public HL7Test()
+        {
+            var path = Path.GetDirectoryName(typeof(HL7Test).GetTypeInfo().Assembly.Location) + "/";
+            this.HL7_ORM = File.ReadAllText(path + "Sample-ORM.txt");
+            this.HL7_ADT = File.ReadAllText(path + "Sample-ADT.txt");
+        }
+
+        [Fact]
+        public void SmokeTest()
+        {
+            Message message = new Message(this.HL7_ORM);
+            Assert.NotNull(message);
+
+            // message.ParseMessage();
+            // File.WriteAllText("SmokeTestResult.txt", message.SerializeMessage(false));
+        }
+
+        [Fact]
+        public void ParseTest1()
+        {
+            var message = new Message(this.HL7_ORM);
+
+            var isParsed = message.ParseMessage();
+            Assert.True(isParsed);
+        }
+
+        [Fact]
+        public void ParseTest2()
+        {
+            var message = new Message(this.HL7_ADT);
+
+            var isParsed = message.ParseMessage();
+            Assert.True(isParsed);
+        }
+
+
+        [Fact]
+        public void ReadSegmentTest()
+        {
+            var message = new Message(this.HL7_ORM);
+            message.ParseMessage();
+
+            Segment MSH_1 = message.Segments("MSH")[0];
+            Assert.NotNull(MSH_1);
+        }
+
+        [Fact]
+        public void ReadDefaultSegmentTest()
+        {
+            var message = new Message(this.HL7_ADT);
+            message.ParseMessage();
+
+            Segment MSH = message.DefaultSegment("MSH");
+            Assert.NotNull(MSH);
+        }
+
+        [Fact]
+        public void ReadFieldTest()
+        {
+            var message = new Message(this.HL7_ADT);
+            message.ParseMessage();
+
+            var MSH_9 = message.GetValue("MSH.9");
+            Assert.Equal("ADT^O01", MSH_9);
+        }
+
+        [Fact]
+        public void ReadComponentTest()
+        {
+            var message = new Message(this.HL7_ADT);
+            message.ParseMessage();
+
+            var MSH_9_1 = message.GetValue("MSH.9.1");
+            Assert.Equal("ADT", MSH_9_1);
+        }
+
+        [Fact]
+        public void AddComponentsTest()
+        {
+            var encoding = new HL7Encoding();
+
+            //Create a Segment with name ZIB
+            Segment newSeg = new Segment("ZIB", encoding);
+
+            // Create Field ZIB_1
+            Field ZIB_1 = new Field("ZIB1", encoding);
+            // Create Field ZIB_5
+            Field ZIB_5 = new Field("ZIB5", encoding);
+
+            // Create Component ZIB.5.3
+            Component com1 = new Component("ZIB.5.3_", encoding);
+
+            // Add Component ZIB.5.3 to Field ZIB_5
+            ZIB_5.AddNewComponent(com1, 3);
+
+            // Overwrite the same field again
+            ZIB_5.AddNewComponent(new Component("ZIB.5.3", encoding), 3);
+
+            // Add Field ZIB_1 to segment ZIB, this will add a new filed to next field location, in this case first field
+            newSeg.AddNewField(ZIB_1);
+
+            // Add Field ZIB_5 to segment ZIB, this will add a new filed as 5th field of segment
+            newSeg.AddNewField(ZIB_5, 5);
+
+            // Add segment ZIB to message
+            var message = new Message(this.HL7_ADT);
+            message.AddNewSegment(newSeg);
+
+            string serializedMessage = message.SerializeMessage(false);
+            Assert.Equal("ZIB|ZIB1||||ZIB5^^ZIB.5.3\r", serializedMessage);
+        }
+
+        [Fact]
+        public void EmptyFieldsTest()
+        {
+            var message = new Message(this.HL7_ADT);
+            message.ParseMessage();
+
+            var NK1 = message.DefaultSegment("NK1").GetAllFields();
+            Assert.Equal(34, NK1.Count);
+            Assert.Equal(string.Empty, NK1[33].Value);
+        }
+
+        [Fact]
+        public void EncodingForOutputTest()
+        {
+            const string oruUrl = "domain.com/resource.html?Action=1&ID=2";  // Text with special character (&)
+
+            var obx = new Segment("OBX", new HL7Encoding());
+            obx.AddNewField("1");
+            obx.AddNewField("RP");
+            obx.AddNewField("70030^Radiologic Exam, Eye, Detection, FB^CDIRadCodes");
+            obx.AddNewField("1");
+            obx.AddNewField(obx.Encoding.Encode(oruUrl));  // Encoded field
+            obx.AddNewField("F", 11);
+            obx.AddNewField(MessageHelper.LongDateWithFractionOfSecond(DateTime.Now), 14);
+
+            var oru = new Message();
+            oru.AddNewSegment(obx);
+
+            var str = oru.SerializeMessage(false);
+
+            Assert.DoesNotContain("&", str);  // Should have \T\ instead
+        }
+
+        [Fact]
+        public void AddFieldTest()
+        {
+            var enc = new HL7Encoding();
+            Segment PID = new Segment("PID", enc);
+            // Creates a new Field
+            PID.AddNewField("1", 1);
+
+            // Overwrites the old Field
+            PID.AddNewField("2", 1);
+
+            Message message = new Message();
+            message.AddNewSegment(PID);
+            var str = message.SerializeMessage(false);
+
+            Assert.Equal("PID|2\r", str);
+        }
+
+        [Fact]
+        public void GetMSH1Test()
+        {
+            var message = new Message(this.HL7_ADT);
+            message.ParseMessage();
+
+            var MSH_1 = message.GetValue("MSH.1");
+            Assert.Equal("|", MSH_1);
+        }
+
+        [Fact]
+        public void GetAckTest()
+        {
+            var message = new Message(this.HL7_ADT);
+            message.ParseMessage();
+            var ack = message.GetACK();
+
+            var MSH_3 = message.GetValue("MSH.3");
+            var MSH_4 = message.GetValue("MSH.4");
+            var MSH_5 = message.GetValue("MSH.5");
+            var MSH_6 = message.GetValue("MSH.6");
+            var MSH_3_A = ack.GetValue("MSH.3");
+            var MSH_4_A = ack.GetValue("MSH.4");
+            var MSH_5_A = ack.GetValue("MSH.5");
+            var MSH_6_A = ack.GetValue("MSH.6");
+
+            Assert.Equal(MSH_3, MSH_5_A);
+            Assert.Equal(MSH_4, MSH_6_A);
+            Assert.Equal(MSH_5, MSH_3_A);
+            Assert.Equal(MSH_6, MSH_4_A);
+
+            var MSH_10 = message.GetValue("MSH.10");
+            var MSH_10_A = ack.GetValue("MSH.10");
+            var MSA_1_1 = ack.GetValue("MSA.1");
+            var MSA_1_2 = ack.GetValue("MSA.2");
+
+            Assert.Equal("AA", MSA_1_1);
+            Assert.Equal(MSH_10, MSH_10_A);
+            Assert.Equal(MSH_10, MSA_1_2);
+        }
+
+        [Fact]
+        public void AddSegmentMSHTest()
+        {
+            var message = new Message();
+            message.AddSegmentMSH("test", "sendingFacility", "test", "test", "test", "ADR^A19", "test", "D", "2.5");
+        }
+
+        [Fact]
+        public void GetNackTest()
+        {
+            var message = new Message(this.HL7_ADT);
+            message.ParseMessage();
+
+            var error = "Error message";
+            var code = "AR";
+            var ack = message.GetNACK(code, error);
+
+            var MSH_3 = message.GetValue("MSH.3");
+            var MSH_4 = message.GetValue("MSH.4");
+            var MSH_5 = message.GetValue("MSH.5");
+            var MSH_6 = message.GetValue("MSH.6");
+            var MSH_3_A = ack.GetValue("MSH.3");
+            var MSH_4_A = ack.GetValue("MSH.4");
+            var MSH_5_A = ack.GetValue("MSH.5");
+            var MSH_6_A = ack.GetValue("MSH.6");
+
+            Assert.Equal(MSH_3, MSH_5_A);
+            Assert.Equal(MSH_4, MSH_6_A);
+            Assert.Equal(MSH_5, MSH_3_A);
+            Assert.Equal(MSH_6, MSH_4_A);
+
+            var MSH_10 = message.GetValue("MSH.10");
+            var MSH_10_A = ack.GetValue("MSH.10");
+            var MSA_1_1 = ack.GetValue("MSA.1");
+            var MSA_1_2 = ack.GetValue("MSA.2");
+            var MSA_1_3 = ack.GetValue("MSA.3");
+
+            Assert.Equal(MSH_10, MSH_10_A);
+            Assert.Equal(MSH_10, MSA_1_2);
+            Assert.Equal(MSA_1_1, code);
+            Assert.Equal(MSA_1_3, error);
+        }
+
+        [Fact]
+        public void EmptyAndNullFieldsTest()
+        {
+            const string sampleMessage = "MSH|^~\\&|SA|SF|RA|RF|20110613083617||ADT^A04|123|P|2.7||||\r\nEVN|A04|20110613083617||\"\"\r\n";
+
+            var message = new Message(sampleMessage);
+            var isParsed = message.ParseMessage();
+            Assert.True(isParsed);
+            Assert.True(message.SegmentCount > 0);
+            var evn = message.Segments("EVN")[0];
+            var expectEmpty = evn.Fields(3).Value;
+            Assert.Equal(string.Empty, expectEmpty);
+            var expectNull = evn.Fields(4).Value;
+            Assert.Null(expectNull);
+        }
+
+        [Fact]
+        public void MessageWithNullsIsReversable()
+        {
+            const string sampleMessage = "MSH|^~\\&|SA|SF|RA|RF|20110613083617||ADT^A04|123|P|2.7||||\r\nEVN|A04|20110613083617||\"\"\r\n";
+            var message = new Message(sampleMessage);
+            message.ParseMessage();
+            var serialized = message.SerializeMessage(false);
+            Assert.Equal(sampleMessage, serialized);
+        }
+
+        [Fact]
+        public void RemoveSegment()
+        {
+            var message = new Message(this.HL7_ADT);
+            message.ParseMessage();
+            Assert.Equal(2, message.Segments("NK1").Count);
+            message.RemoveSegment("NK1", 1);
+            Assert.Single(message.Segments("NK1"));
+            message.RemoveSegment("NK1");
+            Assert.Empty(message.Segments("NK1"));
+        }
+    }
+}
